Changed the dynamic redirect and Id-as-string implementation from Josh.
Dynamic redirect uri changes:
The implementation is now more 'Spring Security'-y by storing the original host name in a custom WebAuthenticationDetails implementation.
Updated the OAuth2ServiceProperties to store the redirectURI as a URI rather than as a String as this makes absolute URI checks easier.

Id type changes:
Changed the OAuth2UserDetailsLoader interface to allow the type of the Id to be specified using Generics. This means that implementations of OAuth2UserDetailsLoader don't have to cast from String to the type they want to use.
A Spring Converter can be optionally specified to convert to the required type.